Split the monolith app.py file into multiple files
